working completions for monolith
