+++ examples/python_rs/llm/tensorrt_llm/monolith/worker.py
@@ -16,23 +16,15 @@
 
 import asyncio
 import signal
-import threading
 import uuid
-from contextlib import asynccontextmanager
 from enum import Enum
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Any, Dict, List, Tuple
 
 import uvloop
+from common.base_engine import BaseTensorrtLLMEngine
 from common.parser import parse_tensorrt_llm_args
-from common.processor import (
-    ChatProcessor,
-    CompletionsProcessor,
-    merge_promises,
-    parse_chat_message_content,
-)
+from common.processor import merge_promises, parse_chat_message_content
 from common.protocol import nvChatCompletionRequest, nvCompletionRequest
-from tensorrt_llm._torch import LLM
-from tensorrt_llm._torch.pyexecutor.config import PyTorchConfig
 from tensorrt_llm.executor import CppExecutorError
 from tensorrt_llm.llmapi.llm import RequestOutput
 from tensorrt_llm.logger import logger
@@ -40,7 +32,6 @@ from tensorrt_llm.serve.openai_protocol import (
     ChatCompletionStreamResponse,
     CompletionStreamResponse,
 )
-from transformers import AutoTokenizer
 
 from triton_distributed.runtime import (
     DistributedRuntime,
@@ -56,110 +47,13 @@ class RequestType(Enum):
     COMPLETION = "completion"
 
 
-class TensorrtLLMEngine:
+class TensorrtLLMEngine(BaseTensorrtLLMEngine):
     """
     Request handler for the generate endpoint
     """
 
     def __init__(self, engine_args: Tuple[Dict[str, Any], Dict[str, Any]]):
-        self.pytorch_config_args, self.llm_engine_args = engine_args
-        self._init_engine()
-        self.model = self.llm_engine_args["model"]
-        if "tokenizer" in self.llm_engine_args.keys():
-            tokenizer = self.llm_engine_args["tokenizer"]
-            self.tokenizer = AutoTokenizer.from_pretrained(tokenizer)
-        else:
-            self.tokenizer = AutoTokenizer.from_pretrained(self.model)
-        self.chat_processor = ChatProcessor(self.model, self.tokenizer)
-        self.completions_processor = CompletionsProcessor(self.model)
-
-    def _init_engine(self):
-        logger.info("Initializing engine")
-        # Run the engine in a separate thread running the AsyncIO event loop.
-        self._llm_engine: Optional[Any] = None
-        self._llm_engine_start_cv = threading.Condition()
-        self._llm_engine_shutdown_event = asyncio.Event()
-        self._event_thread = threading.Thread(
-            target=asyncio.run, args=(self._run_llm_engine(),)
-        )
-        self._event_thread.start()
-        with self._llm_engine_start_cv:
-            while self._llm_engine is None:
-                self._llm_engine_start_cv.wait()
-
-        # The 'threading.Thread()' will not raise the exception here should the engine
-        # failed to start, so the exception is passed back via the engine variable.
-        if isinstance(self._llm_engine, Exception):
-            e = self._llm_engine
-            logger.error(f"Failed to start engine: {e}")
-            if self._event_thread is not None:
-                self._event_thread.join()
-                self._event_thread = None
-            raise e
-
-    async def _run_llm_engine(self):
-        # Counter to keep track of ongoing request counts.
-        self._ongoing_request_count = 0
-
-        @asynccontextmanager
-        async def async_llm_wrapper():
-            # Create LLM in a thread to avoid blocking
-            loop = asyncio.get_running_loop()
-            try:
-                pytorch_config = PyTorchConfig(**self.pytorch_config_args)
-                llm = await loop.run_in_executor(
-                    None,
-                    lambda: LLM(
-                        **self.llm_engine_args, pytorch_backend_config=pytorch_config
-                    ),
-                )
-                yield llm
-            finally:
-                if "llm" in locals():
-                    # Run shutdown in a thread to avoid blocking
-                    await loop.run_in_executor(None, llm.shutdown)
-
-        try:
-            async with async_llm_wrapper() as engine:
-                # Capture the engine event loop and make it visible to other threads.
-                self._event_loop = asyncio.get_running_loop()
-
-                # Signal the engine is started and make it visible to other threads.
-                with self._llm_engine_start_cv:
-                    self._llm_engine = engine
-                    self._llm_engine_start_cv.notify_all()
-
-                logger.info("Engine loaded and ready to serve...")
-
-                # Wait for the engine shutdown signal.
-                await self._llm_engine_shutdown_event.wait()
-
-                # Wait for the ongoing requests to complete.
-                while self._ongoing_request_count > 0:
-                    logger.info(
-                        "Awaiting remaining {} requests".format(
-                            self._ongoing_request_count
-                        )
-                    )
-                    await asyncio.sleep(1)
-
-                # Cancel all tasks in the event loop.
-                for task in asyncio.all_tasks(loop=self._event_loop):
-                    if task is not asyncio.current_task():
-                        task.cancel()
-
-        except Exception as e:
-            # Signal and pass the exception back via the engine variable if the engine
-            # failed to start. If the engine has started, re-raise the exception.
-            with self._llm_engine_start_cv:
-                if self._llm_engine is None:
-                    self._llm_engine = e
-                    self._llm_engine_start_cv.notify_all()
-                    return
-            raise e
-
-        self._llm_engine = None
-        logger.info("Shutdown complete")
+        super().__init__(engine_args)
 
     @triton_endpoint(nvChatCompletionRequest, ChatCompletionStreamResponse)
     async def generate_chat(self, request):
@@ -189,21 +83,20 @@ class TensorrtLLMEngine:
             )
             sampling_params = request.to_sampling_params()
 
-            if self._llm_engine is not None:
-                promise = self._llm_engine.generate_async(
-                    prompt,
-                    sampling_params,
-                    streaming=request.stream,
+            promise = self._llm_engine.generate_async(
+                prompt,
+                sampling_params,
+                streaming=request.stream,
+            )
+            if request.stream:
+                response_generator = self.chat_processor.stream_response(
+                    request, request_id, conversation, promise
                 )
-                if request.stream:
-                    response_generator = self.chat_processor.stream_response(
-                        request, request_id, conversation, promise
-                    )
-                    async for response in response_generator:
-                        yield response
-                else:
-                    # TODO: Implement non-streaming chat completion
-                    raise RuntimeError("Non-streaming is not supported")
+                async for response in response_generator:
+                    yield response
+            else:
+                # TODO: Implement non-streaming chat completion
+                raise RuntimeError("Non-streaming is not supported")
         except CppExecutorError:
             # If internal executor error is raised, shutdown the server
             signal.raise_signal(signal.SIGINT)
@@ -218,8 +111,6 @@ class TensorrtLLMEngine:
         self._ongoing_request_count += 1
         logger.debug(f"Received completion request: {request}")
 
-        if isinstance(request, str):
-            request = nvCompletionRequest.parse_raw(request)
         if isinstance(request.prompt, str) or (
             isinstance(request.prompt, list) and isinstance(request.prompt[0], int)
         ):
@@ -249,12 +140,14 @@ class TensorrtLLMEngine:
             async for response in response_generator:
                 yield response
         else:
-            # TODO: Implement non-streaming completion
-            # response = await self.chat_processor.create_completion_response(
-            #    request, generator, num_choices
+            # TODO: why doesn't it read stream from input?
+            # stream is always True
+
+            raise NotImplementedError("Non-streaming is not supported")
+            # response = await self.completions_processor.create_completion_response(
+            #     request, generator, num_choices
             # )
             # yield response
-            raise RuntimeError("Non-streaming is not supported")
 
 
 @triton_worker()
